// portafolio_AF/src/app/model/education.model.ts
export class Education {
    id?: number;
    eduI: string;
    eduDI: string;
    ediDF: string;
    eduT: string;
    eduDesc: string;

    constructor(eduI: string, eduDI: string, ediDF: string, eduT: string, eduDesc: string) {
        this.eduI = eduI;
        this.eduDI = eduDI;
        this.ediDF = ediDF;
        this.eduT = eduT;
        this.eduDesc = eduDesc;
    }





}